Restructure server code
